# Add lazy var evaluation for reclaim labels