--- a/lua/ui/game/reclaim.lua
+++ b/lua/ui/game/reclaim.lua
@@ -1,20 +1,40 @@
+-- # Imports
+
 local LayoutHelpers = import('/lua/maui/layouthelpers.lua')
 local Group = import('/lua/maui/group.lua').Group
 local Bitmap = import('/lua/maui/bitmap.lua').Bitmap
 local UIUtil = import('/lua/ui/uiutil.lua')
 local Prefs = import('/lua/user/prefs.lua')
 local options = Prefs.GetFromCurrentProfile('options')
+local LazyVar = import('/lua/lazyvar.lua')
 
-local MaxLabels = options.maximum_reclaim_count or 1000 -- The maximum number of labels created in a game session
+-- # Settings
+
+local MaxLabels = 1000 -- The maximum number of labels created in a game session
 local MinAmount = options.minimum_reclaim_amount or 10
 
-local Reclaim = {} -- int indexed list, sorted by mass, of all props that can show a label currently in the sim
-local LabelPool = {} -- Stores labels up too MaxLabels
+-- # Lazy evaluation
+
+local RootOfLabels = false
+local LazyView = LazyVar.Create(false)
+
+-- # internal state
+
+local Thread = false
+
+local Reclaim = { } -- int indexed list, sorted by mass, of all props that can show a label currently in the sim
+local LabelPool = { } -- Stores labels up too MaxLabels
 local OldZoom
 local OldPosition
 local ReclaimChanged = true
 local PlayableArea
 local OutsidePlayableAreaReclaim = {}
+
+-- # Debug properties
+
+local ReclaimLabelsMade = 0
+local UpdateLeft = 0
+local UpdateTop = 0
 
 -- Stores/updates a reclaim entity's data using EntityId as key
 -- called from /lua/UserSync.lua
@@ -73,7 +93,7 @@ function updateMaxLabels(value)
 end
 
 function OnScreen(view, pos)
-    local proj = view:Project(Vector(pos[1], pos[2], pos[3]))
+    local proj = view:Project(pos)
     return not (proj.x < 0 or proj.y < 0 or proj.x > view.Width() or proj.y > view:Height())
 end
 
@@ -84,83 +104,146 @@ function InPlayableArea(pos)
     return true
 end
 
-local WorldLabel = Class(Group) {
-    __init = function(self, parent, position)
+local RootLabel = Class(Group) {
+    __init = function(self, parent, view, camera)
         Group.__init(self, parent)
-        self.parent = parent
-        self.proj = nil
-        self:SetPosition(position)
 
+        -- Allows us to keep track of updating
+        self.View = view
+        self.Camera = camera
+        self.OldCameraPosition = camera:GetFocusPosition()
+        self.OldCameraZoom = camera:GetZoom()
+
+        -- default values
         self.Top:Set(0)
         self.Left:Set(0)
-        LayoutHelpers.SetDimensions(self, 25, 25)
-        self:SetNeedsFrameUpdate(true)
+        self.Width:Set(1)
+        self.Height:Set(1)
     end,
 
-    Update = function(self)
+    -- stop updating when we're hidden
+    OnHide = function (self, hidden)
+        self:SetNeedsFrameUpdate(not hidden)
     end,
 
-    SetPosition = function(self, position)
-        self.position = position or {}
-    end,
+    -- update all labels 
+    OnFrame = function(self)
+        local zoom = self.Camera:GetZoom()
+        local position = self.Camera:GetFocusPosition()
 
-    OnFrame = function(self, delta)
-        self:Update()
-    end
+        local update = false
+        update = update or (zoom ~= self.OldCameraZoom)
+        update = update or (position[1] ~= self.OldCameraPosition[1])
+        update = update or (position[2] ~= self.OldCameraPosition[2])
+        update = update or (position[3] ~= self.OldCameraPosition[3])
+
+        if update then 
+            LazyView:Set(self.View)
+
+            self.OldCameraZoom = zoom 
+            self.OldCameraPosition = position
+        end
+    end,
+}
+
+local Label = Class(Group) {
+    __init = function(self, parent)
+        Group.__init(self, parent)
+        self.parent = parent
+
+        -- default values
+        self.Top:Set(0)
+        self.Left:Set(0)
+        self.Width:Set(25)
+        self.Height:Set(25)
+    end,
 }
 
 -- Creates an empty reclaim label
-function CreateReclaimLabel(view)
-    local label = WorldLabel(view, Vector(0, 0, 0))
+function CreateReclaimLabel(root)
 
+    ReclaimLabelsMade = ReclaimLabelsMade + 1
+    local label = Label(root)
+
+    -- mass bitmap
     label.mass = Bitmap(label)
     label.mass:SetTexture(UIUtil.UIFile('/game/build-ui/icon-mass_bmp.dds'))
     LayoutHelpers.AtLeftIn(label.mass, label)
     LayoutHelpers.AtVerticalCenterIn(label.mass, label)
     LayoutHelpers.SetDimensions(label.mass, 14, 14)
 
+    -- text information
     label.text = UIUtil.CreateText(label, "", 10, UIUtil.bodyFont)
     label.text:SetColor('ffc7ff8f')
     label.text:SetDropShadow(true)
     LayoutHelpers.AtLeftIn(label.text, label, 16)
     LayoutHelpers.AtVerticalCenterIn(label.text, label)
 
+    -- disable various settings
     label:DisableHitTest(true)
-    label.OnHide = function(self, hidden)
-        self:SetNeedsFrameUpdate(not hidden)
-    end
 
-    label.Update = function(self)
-        local view = self.parent.view
-        local proj = view:Project(self.position)
-        LayoutHelpers.AtLeftTopIn(self, self.parent, (proj.x - self.Width() / 2) / LayoutHelpers.GetPixelScaleFactor(), (proj.y - self.Height() / 2 + 1) / LayoutHelpers.GetPixelScaleFactor())
-        self.proj = {x=proj.x, y=proj.y }
+    -- display properties
+    label.DisplayReclaim = function(self, label)
 
-    end
-
-    label.DisplayReclaim = function(self, r)
+        -- show us 
         if self:IsHidden() then
             self:Show()
         end
-        self:SetPosition(r.position)
-        if r.mass ~= self.oldMass then
-            local mass = tostring(math.floor(0.5 + r.mass))
+
+        -- change our position
+        self.Position = label.position
+
+        -- update mass
+        if label.mass ~= self.oldMass then
+            local mass = tostring(math.floor(0.5 + label.mass))
             self.text:SetText(mass)
-            self.oldMass = r.mass
+            self.oldMass = label.mass
         end
     end
 
-    label:Update()
+    -- update left position
+    label.Left:Set(
+        function()
+            local view = LazyView()
+
+            if view then 
+                UpdateLeft = UpdateLeft + 1
+                local projected = view:Project(label.Position)
+                local value = (projected.x - 12) / LayoutHelpers.GetPixelScaleFactor()
+                return value
+            else 
+                return 0
+            end
+        end
+    )
+
+    -- update right position
+    label.Top:Set(
+        function()
+            local view = LazyView()
+
+            if view then 
+                UpdateTop = UpdateTop + 1
+                local projected = view:Project(label.Position)
+                local value = (projected.y - 13) / LayoutHelpers.GetPixelScaleFactor()
+                return value
+            else 
+                return 0 
+            end
+        end
+    )
 
     return label
 end
 
-function UpdateLabels()
+local onScreenReclaims = {}
+function UpdateLabels(root)
+
     local view = import('/lua/ui/game/worldview.lua').viewLeft -- Left screen's camera
 
     local onScreenReclaimIndex = 1
-    local onScreenReclaims = {}
 
+    once = false
     -- One might be tempted to use a binary insert; however, tests have shown that it takes about 140x more time
     for _, r in Reclaim do
         r.onScreen = OnScreen(view, r.position)
@@ -170,7 +253,7 @@ function UpdateLabels()
         end
     end
 
-    table.sort(onScreenReclaims, function(a, b) return a.mass > b.mass end)
+    -- table.sort(onScreenReclaims, function(a, b) return a.mass > b.mass end)
 
     -- Create/Update as many reclaim labels as we need
     local labelIndex = 1
@@ -183,7 +266,7 @@ function UpdateLabels()
             label = nil
         end
         if not label then
-            label = CreateReclaimLabel(view.ReclaimGroup, r)
+            label = CreateReclaimLabel(root, r)
             LabelPool[labelIndex] = label
         end
 
@@ -204,89 +287,86 @@ function UpdateLabels()
     end
 end
 
-local ReclaimThread
 function ShowReclaim(show)
-    local view = import('/lua/ui/game/worldview.lua').viewLeft
-    view.ShowingReclaim = show
-
-    if show and not view.ReclaimThread then
-        view.ReclaimThread = ForkThread(ShowReclaimThread)
+    if show then 
+        if RootOfLabels then 
+            RootOfLabels:Show()
+        end
+    else 
+        if RootOfLabels then 
+            RootOfLabels:Hide()
+        end
     end
-end
-
-function InitReclaimGroup(view)
-    if not view.ReclaimGroup or IsDestroyed(view.ReclaimGroup) then
-        local rgroup = Group(view)
-        rgroup.view = view
-        rgroup:DisableHitTest()
-        LayoutHelpers.FillParent(rgroup, view)
-        rgroup:Show()
-
-        view.ReclaimGroup = rgroup
-        rgroup:SetNeedsFrameUpdate(true)
-    else
-        view.ReclaimGroup:Show()
-    end
-
-    view.NewViewing = true
 end
 
 function ShowReclaimThread(watch_key)
     local view = import('/lua/ui/game/worldview.lua').viewLeft
     local camera = GetCamera("WorldCamera")
 
-    InitReclaimGroup(view)
+    LOG("ShowReclaimThread")
+    if not RootOfLabels then 
+        RootOfLabels = RootLabel(GetFrame(0), view, camera)
+        RootOfLabels.Left:Set(0)
+        RootOfLabels.Top:Set(0)
+        RootOfLabels.Width:Set(1)
+        RootOfLabels.Height:Set(1)
+        RootOfLabels:Show()
+    end
 
-    while view.ShowingReclaim and (not watch_key or IsKeyDown(watch_key)) do
-        if not IsDestroyed(camera) then
-            local zoom = camera:GetZoom()
-            local position = camera:GetFocusPosition()
-            if ReclaimChanged
-                or view.NewViewing
-                or OldZoom ~= zoom
-                or OldPosition[1] ~= position[1]
-                or OldPosition[2] ~= position[2]
-                or OldPosition[3] ~= position[3] then
-                    UpdateLabels()
-                    OldZoom = zoom
-                    OldPosition = position
-                    ReclaimChanged = false
-            end
+    while true do
+        local zoom = camera:GetZoom()
+        local position = camera:GetFocusPosition()
+        if ReclaimChanged
+            or OldZoom ~= zoom
+            or OldPosition[1] ~= position[1]
+            or OldPosition[2] ~= position[2]
+            or OldPosition[3] ~= position[3] then
 
-            view.NewViewing = false
+                -- update labels with regard to which ones we show
+                UpdateLabels(RootOfLabels)
+
+                OldZoom = zoom
+                OldPosition = position
+                ReclaimChanged = false
         end
         WaitSeconds(.1)
     end
-
-    if not IsDestroyed(view) then
-        view.ReclaimThread = nil
-        view.ReclaimGroup:Hide()
-    end
 end
 
-function ToggleReclaim()
-    local view = import('/lua/ui/game/worldview.lua').viewLeft
-    ShowReclaim(not view.ShowingReclaim)
-end
+-- function ToggleReclaim()
+--     local view = import('/lua/ui/game/worldview.lua').viewLeft
+--     ShowReclaim(not view.ShowingReclaim)
+-- end
 
 -- Called from commandgraph.lua:OnCommandGraphShow()
 local CommandGraphActive = false
 function OnCommandGraphShow(bool)
-    local view = import('/lua/ui/game/worldview.lua').viewLeft
-    if view.ShowingReclaim and not CommandGraphActive then return end -- if on by toggle key
 
     CommandGraphActive = bool
     if CommandGraphActive then
         ForkThread(function()
             local keydown
             while CommandGraphActive do
+
                 keydown = IsKeyDown('Control')
-                if keydown ~= view.ShowingReclaim then -- state has changed
-                    ShowReclaim(keydown)
+                LOG(keydown)
+                if keydown then 
+                    ShowReclaim(true)
+                    
+                    if not Thread then 
+                        LOG("Creating reclaim thread!")
+                        Thread = ForkThread(ShowReclaimThread)
+                    end
                 end
+
+                LOG("UpdateLeft: " .. tostring(UpdateLeft))
+                LOG("UpdateTop: " .. tostring(UpdateTop))
+                LOG("ReclaimLabelsMade: " .. tostring(ReclaimLabelsMade))
+
                 WaitSeconds(.1)
             end
 
+            KillThread(Thread)
             ShowReclaim(false)
         end)
     else