This is an experimental branch with regard to reclaim labels - it will be embedded into #3534 .

I noticed that these labels were not using lazy evaluation, and instead re-creating functions and closures each frame to update their positions. On this branch they use lazy evaluation. This works like a charm because now the labels are only updated during movement and not during each individual frame. This has a significant performance boost: just rendering the 1000 labels is not an issue. But, updating them (even through lazy evaluation) the memory usage just spikes (~10 mb per second) and then it stutters because the garbage collector kicks in. 

@Crotalus You mentioned using lazy evaluation as a way to speed up the performance of the UI. Do you have any idea what causes the massive memory usage?

Steps to reproduce:
 - Turn on `ShowStats` in the console
 - Open up `Heap` and `Frame` 
 - Start the map `Forest Nothing`
 - Zoom out after starting, then press CTRL + SHIFT to show reclaim
 - Watch framerate drop when moving, but stable when not moving and check logs to see how often things are updated

I have no idea where the memory is being used for. At least when merging with the other PR the performance will be steady when the camera is not moving at all.

And another question: how does the lazy evaluation work internally? I found the file in the base game, but it isn't quite as informative.