@Garanas 2021-11-14
The performance status at this point, with a base line fps of 90 when no labels are shown:
 - When showing 1000 labels without moving the camera: 15 fps -> 80 fps
 - When showing 1000 labels while moving the camera: 15 fps -> 35 fps
 
 There's also the question of grouping labels: is there interesting in this? I personally think that having a 1000 labels doesn't make anything easier. Perhaps 100 - 150 tops and combine them when the number of labels becomes too high.